Move function limiters into silo.base and fix mistakes
