+++ src/sentry/integrations/gitlab/client.py
@@ -13,10 +13,9 @@ from sentry.integrations.gitlab.utils import GitLabApiClientPath
 from sentry.integrations.mixins.commit_context import FileBlameInfo, SourceLineInfo
 from sentry.models.repository import Repository
 from sentry.services.hybrid_cloud.identity.model import RpcIdentity
-from sentry.services.hybrid_cloud.util import control_silo_function
 from sentry.shared_integrations.client.proxy import IntegrationProxyClient
 from sentry.shared_integrations.exceptions import ApiError, ApiUnauthorized
-from sentry.silo.base import SiloMode
+from sentry.silo.base import SiloMode, control_silo_function
 from sentry.utils import metrics
 from sentry.utils.http import absolute_uri
