+++ src/sentry/integrations/msteams/client.py
@@ -10,9 +10,8 @@ from sentry.integrations.client import ApiClient
 from sentry.models.integrations import Integration
 from sentry.services.hybrid_cloud.integration import integration_service
 from sentry.services.hybrid_cloud.integration.model import RpcIntegration
-from sentry.services.hybrid_cloud.util import control_silo_function
 from sentry.shared_integrations.client.proxy import IntegrationProxyClient, infer_org_integration
-from sentry.silo.base import SiloMode
+from sentry.silo.base import SiloMode, control_silo_function
 
 # five minutes which is industry standard clock skew tolerance
 CLOCK_SKEW = 60 * 5
