+++ src/sentry/integrations/utils/identities.py
@@ -11,7 +11,7 @@ from sentry.models.integrations.organization_integration import OrganizationInte
 from sentry.models.user import User
 from sentry.services.hybrid_cloud.organization import RpcOrganization, organization_service
 from sentry.services.hybrid_cloud.user.service import user_service
-from sentry.services.hybrid_cloud.util import control_silo_function
+from sentry.silo.base import control_silo_function
 
 _logger = logging.getLogger(__name__)
