+++ src/sentry/integrations/utils/sync.py
@@ -11,7 +11,7 @@ from sentry.models.organization import Organization
 from sentry.models.project import Project
 from sentry.services.hybrid_cloud.integration import integration_service
 from sentry.services.hybrid_cloud.user.service import user_service
-from sentry.services.hybrid_cloud.util import region_silo_function
+from sentry.silo.base import region_silo_function
 from sentry.tasks.integrations import sync_assignee_outbound
 
 if TYPE_CHECKING:
