+++ src/sentry/middleware/integrations/parsers/msteams.py
@@ -19,7 +19,7 @@ from sentry.integrations.types import EXTERNAL_PROVIDERS, ExternalProviders
 from sentry.models.integrations.integration import Integration
 from sentry.models.integrations.organization_integration import OrganizationIntegration
 from sentry.models.outbox import WebhookProviderIdentifier
-from sentry.services.hybrid_cloud.util import control_silo_function
+from sentry.silo.base import control_silo_function
 from sentry.types.region import Region, RegionResolutionError
 
 logger = logging.getLogger(__name__)
