+++ src/sentry/middleware/integrations/parsers/vsts.py
@@ -11,7 +11,7 @@ from sentry.integrations.vsts.webhooks import WorkItemWebhook, get_vsts_external
 from sentry.models.integrations.integration import Integration
 from sentry.models.integrations.organization_integration import OrganizationIntegration
 from sentry.models.outbox import WebhookProviderIdentifier
-from sentry.services.hybrid_cloud.util import control_silo_function
+from sentry.silo.base import control_silo_function
 
 logger = logging.getLogger(__name__)
