+++ src/sentry/receivers/core.py
@@ -16,10 +16,9 @@ from sentry.models.project import Project
 from sentry.models.projectkey import ProjectKey
 from sentry.models.team import Team
 from sentry.services.hybrid_cloud.user.service import user_service
-from sentry.services.hybrid_cloud.util import region_silo_function
 from sentry.services.organization import organization_provisioning_service
 from sentry.signals import post_upgrade, project_created
-from sentry.silo.base import SiloMode
+from sentry.silo.base import SiloMode, region_silo_function
 from sentry.utils.env import in_test_environment
 from sentry.utils.settings import is_self_hosted
