+++ src/sentry/rules/actions/integrations/create_ticket/utils.py
@@ -12,8 +12,8 @@ from sentry.models.grouplink import GroupLink
 from sentry.models.integrations.external_issue import ExternalIssue
 from sentry.services.hybrid_cloud.integration.model import RpcIntegration
 from sentry.services.hybrid_cloud.integration.service import integration_service
-from sentry.services.hybrid_cloud.util import region_silo_function
 from sentry.shared_integrations.exceptions import IntegrationFormError
+from sentry.silo.base import region_silo_function
 from sentry.types.rules import RuleFuture
 from sentry.utils import metrics
